add Roi Color option

+++ PreStage/ROI/make_roi.py
@@ -2,12 +2,13 @@ import os
 import cv2
 import numpy as np
 class selectROI():
-    def __init__(self, direc = os.getcwd(), replaceImg = False, scale = 1):
+    def __init__(self, direc = os.getcwd(), replaceImg = False, scale = 1, ROIColor = (0,0,0)):
         self.get_image_list(direc)
         self.direc = direc
         self.replaceImg = replaceImg
         self.scale = scale
         self.roiName = 'ROI.jpg'
+        self.ROIColor = ROIColor
     #Get all image from direc
     def get_image_list(self, direc):
         os.chdir(direc)
@@ -45,13 +46,14 @@ class selectROI():
                 # scale ROI to [0, 1] => binary mask
                 thresh, roi = cv2.threshold(roi, thresh=128, maxval=1, type=cv2.THRESH_BINARY)
                 # apply ROI on the original image
-                img[np.where((roi==[0,0,0]).all(axis=2))] = [0,0,255]
+                img[np.where((roi==[0,0,0]).all(axis=2))] = self.ROIColor
                 savename = folder + path[:-3] + "png"
                 cv2.imwrite(savename, img)
                 print('ROI ' + savename + ' successfull!')
             print('Done!...')
-            print('Remove ' + self.roiName)
-            os.remove(self.roiName)
+            if self.replaceImg:
+                print('Remove ' + self.roiName)
+                os.remove(self.roiName)
 
     #Ve Roi mask    
     def createROI(self):
